bpo-46205: exit if no workers are alive in runtest_mp
There is a race condition in runtest_mp where if a worker already
pushed its final output to the queue, but is still alive, then the
the main thread waits forever on the the now-empty queue.

The workers now indicate they have finished by putting a sentintel
"WorkerExited" object in the output queue.